Build a sword and hit it with a hammer
Controls for movement added.
Left thumbstick is for moving, right is for rotating the camera for the people with neck pain ;) .

Hands and hands animation added.

Hands grabbing added. To grab an object, the object has to have the correct component. After that just reach with your hand and press the grip button (side trigger on the Q2 controller).

Weapon Parts added, Blade, guard and handle place holders that can be moved by the player.

Weapon building added. Currently each part has a red box that shows a snap point (not needed to connect them but recommended to use). Just align the red boxes of the different parts and they will snap to each other. After all the parts have been connected take the hammer and hit the GUARD of the sword to make it.