Started Final Scene + Crafting
Can craft a weapon now.

Go to the smithing room by putting on the VR Headset. Choose what you want to build from the hand menu (Only 1 sword for now). Buy the parts displayed on the crafting bench. Craft the blade (Heat->Hammer x 4 -> Heat -> Hammer x 4 -> Cool in the water - > Sharpen on the robot)
Place each part on the crafting bench and smack it with the hammer. It SHOULD give you a sword.

Some attach points are with wrong rotations but yea to lazy to fix now.